+++ metadata.rb
@@ -1,9 +1,9 @@
-name             'docker-cookbook'
+name             'docker'
 provaides        'docker'
 maintainer       'Gregório Melo'
 maintainer_email 'https://github.com/gregoriomelo'
 license          'All rights reserved'
-description      'Installs docker-cookbook'
+description      'Installs docker'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'