+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-# CHANGELOG for docker-cookbook
+# CHANGELOG for docker cookbook
 
-This file is used to list changes made in each version of docker-cookbook.
+This file is used to list changes made in each version of docker cookbook.
 
 ## 0.1.0:
 
-* Initial release of docker-cookbook
+* Initial release of docker cookbook